crypto_bot/bot/pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
gnews = "^0.3.7"
asyncstdlib = "^3.12.4"
python-dotenv = "^1.0.1"
pymongo = {extras = ["srv"], version = "^4.8.0"}
pytz = "^2024.1"
aiohttp = "^3.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
